(#4) Add semantic version support
This change replaces the validation of the version string using
Gem::Version.new() with a regular expression matching the semver.org
description of versions.

The reason for this change is that Gem::Version.new() is too restrictive
and does not allow semantic version strings to be used.  We're trying to
use semantic versions for modules and for development releases these may
be in the form of 0.0.2git-8-g3d316d1

Also changed the error message to indicate we support semantic versions and
provide the end user a link to semver.org for more information.