(#6) Install modules to their name not "full name"
Without this change, the puppet-module tool installs modules into
directories named of the format author-modulename.  This is problematic
because the Puppet autoloader convention causes Puppet to expect a class
named "mcollective" in a directory named "mcollective" in
manifests/init.pp in the module.

As a result, all modules should be installed into a directory with their
specific module name, NOT a directory prefixed with the author of the
module and a hyphen.

This change replaces the use of the "full_name" with only the
"module_name" instance variables in the Unpacker class.

Reviewed-by: Michael Stanhke <stahnma@puppetlabs.com>