(#6) Add check for symlink when installing
When installing a module, File.exist? returns false for a dangling
symbolic link.  The dangling symlink prevents the actual installation of
the module form succeeding, however, which is the point of checking for
existence of a file with the same name.

This change adds an additional check to see if a symbolic link exists
with the same name of the module we're installing.  Without this change,
a backtrace error was being triggered.

Reviewed-by: Michael Stanhke <stahnma@puppetlabs.com>