(#8506) Add ability to install modules with hyphens
This change fixes the module filename parsing method to support module
names with hyphens and dashes in them.

Without this change the following module name was causing an error in
the install action:

    puppetlabs-cloud-provisioner-0.0.1git-95-g6541187.tar.gz

This error was caused by the module tool doing a simple split on the
filename.  This change uses a more robust regular expression to parse
the username, module name, and semantic version from the package
filename.

Reviewed-by: Dan Bode