(#10438) Add install_dir to the command line
The previous commit for #10438 to have a configurable install dir forgot
to put the command line option in.  This commit does that.

It also adds some code to the spec helper to prevent the specs from
interacting with puppet's actual vardir.  This was necessary to make the
test for the install_dir option workable, and it's just a good thing
besides.